Changed WordStatistic.Appearances to be a long.

+++ src/AtelierTomato.Markov.Model/WordStatistic.cs
@@ -3,6 +3,6 @@
 	public record WordStatistic
 	(
 		string Name,
-		int Appearances
+		long Appearances
 	);
 }
